Added command "rds list".
Added configuration variable OS_USER_DOMAIN_ID. If not set
in environment, it is determined from IAM REST response.
In OTC 1.01, RDS instances are associated to domains instead
of projects. This is a known issue and planned to be changed
to stay compatible with the managment model used in Trove API.