Fix otc images upload to work with new correct API.
image upload via s3 never was openstack compatible, so the file upload
interface should not have been abused. This was corrected by Huawei
and a v1 cloudimages private API is now offered for this.
Use the private API -- if we get an error message (Api does not exist),
try to fall back to the abused interface for backwards compat.