Fix getPortID() by directly doing REST call.
This was broken by getECSVM being fixed to return proper JSON.
Rather than adjusting, do REST call directly.

When we assign an EIP (the old way), we do have an ECS ID, so output it.

Fix bug resulting in not detecting EIP to exist in convertEipToId().

Bump version to 0.7.11 ... but this is not yet 0.7.11 final:
I'd like to rather use publicip: eipid to assign an EIP to a booting VM,
so this will be tested (and implemented if found reasonable) b/f 0.7.11 final.